requirements.txt
streamlit
easyocr
mysql-connector-python